// src/shared/constants/enums/colors.enum.ts
enum FontColors {
    BLUE = '#009FFA',
    BLUE_LIGHT = '#249CF3',
    GRAY = '#A19CAC',
    GRAY_LIGHT = '#A09CAB',
    WHITE = '#fff',
    RED = '#FF757A',
    DARK = '#222',
    DARK_2 = '#151515',
    DARK_LIGHT = '#EFF1F5',
    GRAY_DARK = '#F1F1F1',
    GRAY_DARK_2 = '#C1C1C1',
    BLACK = '#000',
}

export { FontColors };