// src/shared/constants/enums/fontFamilies.ts
enum InterFontFamily {
    ff100 = 'Inter_100Thin',
    ff200 = 'Inter_200ExtraLight',
    ff300 = 'Inter_300Light',
    ff400 = 'Inter_400Regular',
    ff500 = 'Inter_500Medium',
    ff600 = 'Inter_600SemiBold',
    ff700 = 'Inter_700Bold',
    ff800 = 'Inter_800ExtraBold',
    ff900 = 'Inter_900Black',
}

enum RobotoFontFamily {
    ff100 = 'Roboto_100Thin',
    ff300 = 'Roboto_300Light',
    ff400 = 'Roboto_400Regular',
    ff500 = 'Roboto_500Medium',
    ff700 = 'Roboto_700Bold',
    ff900 = 'Roboto_900Black',
}

export { InterFontFamily, RobotoFontFamily };